requirements.txt
facenet_pytorch
flask
torch
torchvision
Pillow
